sampling perf tweaks and make it actually work
To ensure we can immediately sample from *any* cached token position, we
request logits for every decoded token. This is generally quite slow, so
there are some changes to how sampling works.

Consider the case of llama3 with its 128256 vocab size. llama_token_data
is 12 bytes, so each set of logits is 1.46 MiB. For single clients, that
isn't all that much. If batch size is 512, we end up transferring ~751.5
MiB of logit data on every forward pass! That's not happy fun time.

We also have to store all those logits in memory, which is not great for
large cache sizes. A 32K token cache would need ~46.96 GiB of cpu ram!

The primary difference being that we only copy top_k logits from the ptr
given to us by llama.cpp, and sort inline. That's enough to improve perf
of logit ingest by a factor of ~5x on my cpu.

Additionally, this adds custom impls of softmax & truncation sampling to
minimize allocations and operate in place wherever we can. We also store
logit values contiguously to permit auto-vectorization.

All together, the algorithmic improvements are:
- top_k truncation at logit ingest; 8 bytes * top_k space
- in place fused temperature + softmax
- in place fused min-p and p-step truncation
- in place distribution gen for weighted rng sampling

Latency induced from sampling large batches, *and* from ingesting logits
at all positions can now be ignored as it is so irrelevantly tiny.